CVSL-1911: Add comCaseloadService and move all COM related case load to it.
