+++ server/routes/changeTeam/index.ts
@@ -5,7 +5,7 @@ import roleCheckMiddleware from '../../middleware/roleCheckMiddleware'
 import ChangeTeamsLocation from './handlers/changeTeam'
 import { Services } from '../../services'
 
-export default function Index({ conditionService, caseloadService }: Services): Router {
+export default function Index({ conditionService, comCaseloadService }: Services): Router {
   const createPrefix = (path: string) => `/licence/create/caseload${path}`
   const varyPrefix = (path: string) => `/licence/vary/caseload${path}`
 
@@ -21,11 +21,11 @@ export default function Index({ conditionService, caseloadService }: Services): 
       asyncMiddleware(handler)
     )
 
-  const createTeamHandler = new ChangeTeamsLocation(caseloadService, 'create')
+  const createTeamHandler = new ChangeTeamsLocation(comCaseloadService, 'create')
   get(createPrefix('/change-team'), createTeamHandler.GET())
   post(createPrefix('/change-team'), createTeamHandler.POST())
 
-  const varyTeamHandler = new ChangeTeamsLocation(caseloadService, 'vary')
+  const varyTeamHandler = new ChangeTeamsLocation(comCaseloadService, 'vary')
   get(varyPrefix('/change-team'), varyTeamHandler.GET())
   post(varyPrefix('/change-team'), varyTeamHandler.POST())
